feat: add folder views + assets, HTML views
The folders Assets + Views are added in order to standardize the project in comparison to my counterpart in the other branch. Certain aspects of the CSS in Login are re-worked and the login system is fully integrated, working with a JavaScript script. The HTML files for the rest of the views are created and have been given a template. The view PrincipalAlumnos has been finished while work on CursoAlumno is underway.